Move permission level color/lock icon to website extension
