+++ ProjectLighthouse.Servers.Website/Pages/Partials/Links/UserLinkPartial.cshtml
@@ -1,6 +1,6 @@
 @using LBPUnion.ProjectLighthouse.Database
 @using LBPUnion.ProjectLighthouse.Localization
-@using LBPUnion.ProjectLighthouse.Types.Users
+@using LBPUnion.ProjectLighthouse.Servers.Website.Extensions
 @model LBPUnion.ProjectLighthouse.Types.Entities.Profile.UserEntity
 
 @{
