+++ ProjectLighthouse.Servers.Website/Pages/Partials/SlotCardPartial.cshtml
@@ -5,7 +5,6 @@
 @using LBPUnion.ProjectLighthouse.Servers.Website.Extensions
 @using LBPUnion.ProjectLighthouse.Types.Entities.Profile
 @using LBPUnion.ProjectLighthouse.Types.Users
-@using LBPUnion.ProjectLighthouse.Types.Entities.Level
 @using Microsoft.EntityFrameworkCore
 @model LBPUnion.ProjectLighthouse.Types.Entities.Level.SlotEntity
